Stabilize content helper boundaries for release
The content helper had become the last broad runtime bucket after the Chrome-only release work. This keeps the createContentHelpers facade stable for bootstrap and events while moving targeting, extraction, overlay, persistence, and clipboard concerns into focused modules.

The pass also locks the behavior that matters for this split with Playwright coverage around rich target extraction, table TSV export, selection priority, shortcut copy, excluded domains, and reload recovery.

Constraint: Chrome-only contract and dist/chrome remain the release surface

Constraint: No product scope expansion or root-authored runtime JS

Rejected: Rewrite bootstrap/events around new imports | higher blast radius with no user-facing value

Rejected: Keep helpers.js as a large bucket | future edits would keep mixing unrelated runtime responsibilities

Confidence: high

Scope-risk: moderate

Directive: Keep createContentHelpers as the content facade unless bootstrap/events are refactored in the same verified pass

Tested: npm run verify:release

Not-tested: Human manual acceptance in visible Chrome
